% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codon_usage}
\alias{codon_usage}
\title{Codon usage}
\format{
\subsection{codon_usage}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{amino_acid}{Codon sequence in amino acid single letter code}
\item{nucl}{Codon sequence in nucleotide}
\item{percent}{Frequency of codon usage found in human}
}
}
}
\source{
E-mail from Twist.
}
\usage{
codon_usage
}
\description{
A table describing the frequency of each codon in human. Provided by Twist.
}
\keyword{datasets}
