% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_expected_sequences.R
\name{produce_expected_sequences}
\alias{produce_expected_sequences}
\title{Produce list of expected sequences}
\usage{
produce_expected_sequences(reference, codon_offset = 0, no.init.codon = TRUE)
}
\arguments{
\item{reference}{The WT sequence of the region of interest. Must be a
character string containing only A, C, G or T and must have a length that is
a multiple of 3.}

\item{codon_offset}{The offset of the region of interest compared to the
real start site. For example, if the region of interest starts at the 4th
codon of the gene, the \code{codon_offset} would be 3. Default: 0}

\item{no.init.codon}{Should the first codon considered not to be an
initialization codon? Default: \code{TRUE}}
}
\description{
When doing a miniscreen, there is only a small subset of all possible
sequences that are expected. For instance, we do not expect more than 1
codon to be mutated in a typical DMS library.
}
\details{
This function will list all those possible sequences compared to the
provided reference sequence using the 63 alternative codons.
}
\examples{
expected_sequences <- produce_expected_sequences("ACGTGCAAC")

}
